+++ spring-terminator-api/src/test/java/com/dwarfeng/springterminator/api/integration/configuration/ExceptionCodeOffsetConfiguration.java
@@ -1,0 +1,25 @@
+package com.dwarfeng.springterminator.api.integration.configuration;
+
+import com.dwarfeng.springterminator.sdk.util.ServiceExceptionCodes;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Configuration;
+
+import javax.annotation.PostConstruct;
+
+@Configuration
+public class ExceptionCodeOffsetConfiguration {
+
+    @Value("${terminator.exception_code_offset}")
+    private int exceptionCodeOffset;
+    @Value("${terminator.exception_code_offset.subgrade}")
+    private int subgradeExceptionCodeOffset;
+    @Value("${terminator.exception_code_offset.telqos}")
+    private int telqosExceptionCodeOffset;
+
+    @PostConstruct
+    public void init() {
+        ServiceExceptionCodes.setExceptionCodeOffset(exceptionCodeOffset);
+        com.dwarfeng.subgrade.sdk.exception.ServiceExceptionCodes.setExceptionCodeOffset(subgradeExceptionCodeOffset);
+        com.dwarfeng.springtelqos.sdk.util.ServiceExceptionCodes.setExceptionCodeOffset(telqosExceptionCodeOffset);
+    }
+}